Replace stack data structure with Flask sessions
This change is needed to prevent concurrency problems that may
be encountered when more than one user is using the site at the
same time.

The problem will happen when one user's actions cause the stack
value from another user to be popped off. By using sessions we
avoid this problem by storing the genre variable on a per-session
basis.